apply changes to correct file
